Separate eviction logic from preempting queue scheduler
